screen_recorder/requirements.txt
Pillow
piexif
imagehash
pyobjc